Add documentation files
Added `process.md` and `plan.md` to gem. The process file is a log of
my step-by-step actions in creating the gem, including challenges I
faced so far. My goal is to use it for my blog post that accompanies
this gem project. The plan file outlines what I want my app to do --
how it interacts with the users and their commands.