Modify app plan
I had to get rid of the Style A-Z characteristics for each beer. The
way that it needed to be scraped from the website wasn't conducive to
the file setup that I have. I've downloaded a zipfile of my last push
so I have it to go back and work on later if I decide to update my gem.